==> notebooks/04 Deploy Classifier Web Service.ipynb <==
# Converting a Keras model to ONNX

In the steps that follow, you will convert Keras model you just trained to the ONNX format. This will enable you to use this model for classification in a very broad range of environments including:

- Web services
- iOS and Android mobile apps
- Windows apps
- IoT devices

Furthermore, ONNX runtimes and libraries are also designed to maximize performance on some of the best hardware in the industry. In this lab, we will compare the Inference performance of the ONNX vs Keras models.

First we will load the trained Keras model from file, and then convert the model to ONNX.

## Load the Keras Model

Load the saved Keras model. We will convert the Keras model to ONNX format.

import os
import numpy as np
import pandas as pd

np.random.seed(125)

import tensorflow as tf
from tensorflow import keras
from tensorflow.keras.models import load_model
import joblib

output_folder = './outputs'
model_filename = 'final_model.hdf5'

keras_model = load_model(os.path.join(output_folder, model_filename))
print(keras_model.summary())

## Convert to ONNX

Convert the loaded Keras model to ONNX format, and save the ONNX model to the deployment folder.

*Ignore the WARNING - Can't import tf2onnx module, so the conversion on a model with any custom/lambda layer will fail!*

import onnxmltools

deployment_folder = 'deploy'
onnx_export_folder = 'onnx'

# Convert the Keras model to ONNX
onnx_model_name = 'claim_classifier.onnx'
converted_model = onnxmltools.convert_keras(keras_model, onnx_model_name, target_opset=7)

# Save the model locally...
onnx_model_path = os.path.join(deployment_folder, onnx_export_folder)
os.makedirs(onnx_model_path, exist_ok=True)
onnxmltools.utils.save_model(converted_model, os.path.join(onnx_model_path,onnx_model_name))

## Make Inference using the ONNX Model

- Create an ONNX runtime InferenceSession
- Review the expected input shape to make inferences
- Prepare test data
- Make inferences using both the ONNX and the Keras Model on the test data

### ONNX Runtime InferenceSession

import onnxruntime
# Load the ONNX model and observe the expected input shape
onnx_session = onnxruntime.InferenceSession(
    os.path.join(os.path.join(deployment_folder, onnx_export_folder), onnx_model_name))
input_name = onnx_session.get_inputs()[0].name
output_name = onnx_session.get_outputs()[0].name
print('Expected input shape: ', onnx_session.get_inputs()[0].shape)

### Prepare test data

**Load the GloVe word vectors**

word_vectors_dir = './word_vectors'

dictonary = np.load(os.path.join(word_vectors_dir, 'wordsList.npy'))
dictonary = dictonary.tolist()
dictonary = [word.decode('UTF-8') for word in dictonary]
print('Loaded the dictonary! Dictonary size: ', len(dictonary))

word_vectors = np.load(os.path.join(word_vectors_dir, 'wordVectors.npy'))
print ('Loaded the word vectors! Shape of the word vectors: ', word_vectors.shape)

**Create the word contractions map**

contractions_url = ('https://quickstartsws9073123377.blob.core.windows.net/'
                    'azureml-blobstore-0d1c4218-a5f9-418b-bf55-902b65277b85/glove50d/contractions.xlsx')
contractions_df = pd.read_excel(contractions_url)
contractions = dict(zip(contractions_df.original, contractions_df.expanded))

**Setup the helper functions to process the test data**

import re
import string

def remove_special_characters(token):
    pattern = re.compile('[{}]'.format(re.escape(string.punctuation)))
    filtered_token = pattern.sub('', token)
    return filtered_token

def convert_to_indices(corpus, dictonary, c_map, unk_word_index = 399999):
    sequences = []
    for i in range(len(corpus)):
        tokens = corpus[i].split()
        sequence = []
        for word in tokens:
            word = word.lower()
            if word in c_map:
                resolved_words = c_map[word].split()
                for resolved_word in resolved_words:
                    try:
                        word_index = dictonary.index(resolved_word)
                        sequence.append(word_index)
                    except ValueError:
                        sequence.append(unk_word_index) #Vector for unkown words
            else:
                try:
                    clean_word = remove_special_characters(word)
                    if len(clean_word) > 0:
                        word_index = dictonary.index(clean_word)
                        sequence.append(word_index)
                except ValueError:
                    sequence.append(unk_word_index) #Vector for unkown words
        sequences.append(sequence)
    return sequences

**Preprocess the test data**

from tensorflow.keras.preprocessing.sequence import pad_sequences
maxSeqLength = 125

test_claim = ['I crashed my car into a pole.']

test_claim_indices = convert_to_indices(test_claim, dictonary, contractions)
test_data = pad_sequences(test_claim_indices, maxlen=maxSeqLength, padding='pre', truncating='post')

# convert the data type to float
test_data_float = np.reshape(test_data.astype(np.float32), (1,maxSeqLength))

### Make Inferences

Make inferences using both the ONNX and the Keras Model on the test data

# Run an ONNX session to classify the sample.
print('ONNX prediction: ', onnx_session.run([output_name], {input_name : test_data_float}))

# Use Keras to make predictions on the same sample
print('Keras prediction: ', keras_model.predict(test_data_float))

## Compare Inference Performance: ONNX vs Keras

Evaluate the performance of ONNX and Keras by running the same sample 1,000 times. Run the next three cells and compare the performance in your environment.

# Next we will compare the performance of ONNX vs Keras
import timeit
n = 1000

start_time = timeit.default_timer()
for i in range(n):
    keras_model.predict(test_data_float)
keras_elapsed = timeit.default_timer() - start_time
print('Keras performance: ', keras_elapsed)

start_time = timeit.default_timer()
for i in range(n):
    onnx_session.run([output_name], {input_name : test_data_float})
onnx_elapsed = timeit.default_timer() - start_time
print('ONNX performance: ', onnx_elapsed)
print('ONNX is about {} times faster than Keras'.format(round(keras_elapsed/onnx_elapsed)))

# Deploy ONNX model to Azure Kubernetes Services (AKS)


## Connect to an Azure Machine Learning Workspace

import azureml.core

print(azureml.core.VERSION)

from azureml.core.workspace import Workspace

ws = Workspace.from_config()
print(ws)

## Register the model with Azure Machine Learning

In the following, you register the model with Azure Machine Learning (which saves a copy in the cloud).

#Register the model and vectorizer
from azureml.core.model import Model

registered_model_name = 'claim_classifier_onnx'
onnx_model_path = os.path.join(os.path.join(deployment_folder, onnx_export_folder), onnx_model_name)

registered_model = Model.register(model_path = onnx_model_path, # this points to a local file
                       model_name = registered_model_name, # this is the name the model is registered with         
                       description = "Claims classification model.",
                       workspace = ws)

print(registered_model.name, registered_model.description, registered_model.version)

## Create the scoring web service

When deploying models for scoring with Azure Machine Learning services, you need to define the code for a simple web service that will load your model and use it for scoring. By convention this service has two methods init which loads the model and run which scores data using the loaded model.

This scoring service code will later be deployed inside of a specially prepared Docker container.

**Save the scoring web service Python file**

Note that the scoring web service needs the registered model: the ONNX model to make inferences.

%%writefile scoring_service.py
import string
import re
import os
import numpy as np
import pandas as pd
import urllib.request
import json
import tensorflow as tf
from tensorflow import keras
from tensorflow.keras.preprocessing.sequence import pad_sequences
from azureml.core.model import Model
import onnxruntime

def init():

    global onnx_session
    global dictonary
    global contractions
    
    try:
        words_list_url = ('https://quickstartsws9073123377.blob.core.windows.net/'
                          'azureml-blobstore-0d1c4218-a5f9-418b-bf55-902b65277b85/glove50d/wordsList.npy')
        word_vectors_dir = './word_vectors'
        os.makedirs(word_vectors_dir, exist_ok=True)
        urllib.request.urlretrieve(words_list_url, os.path.join(word_vectors_dir, 'wordsList.npy'))
        dictonary = np.load(os.path.join(word_vectors_dir, 'wordsList.npy'))
        dictonary = dictonary.tolist()
        dictonary = [word.decode('UTF-8') for word in dictonary]
        print('Loaded the dictonary! Dictonary size: ', len(dictonary))
        
        contractions_url = ('https://quickstartsws9073123377.blob.core.windows.net/'
                            'azureml-blobstore-0d1c4218-a5f9-418b-bf55-902b65277b85/glove50d/contractions.xlsx')
        contractions_df = pd.read_excel(contractions_url)
        contractions = dict(zip(contractions_df.original, contractions_df.expanded))
        print('Loaded contractions!')
        
        # Retrieve the path to the model file using the model name
        onnx_model_name = 'claim_classifier_onnx'
        onnx_model_path = Model.get_model_path(onnx_model_name)
        print('onnx_model_path: ', onnx_model_path)
        
        onnx_session = onnxruntime.InferenceSession(onnx_model_path)
        print('Onnx Inference Session Created!')
        
    except Exception as e:
        print(e)
        
def remove_special_characters(token):
    pattern = re.compile('[{}]'.format(re.escape(string.punctuation)))
    filtered_token = pattern.sub('', token)
    return filtered_token

def convert_to_indices(corpus, dictonary, c_map, unk_word_index = 399999):
    sequences = []
    for i in range(len(corpus)):
        tokens = corpus[i].split()
        sequence = []
        for word in tokens:
            word = word.lower()
            if word in c_map:
                resolved_words = c_map[word].split()
                for resolved_word in resolved_words:
                    try:
                        word_index = dictonary.index(resolved_word)
                        sequence.append(word_index)
                    except ValueError:
                        sequence.append(unk_word_index) #Vector for unkown words
            else:
                try:
                    clean_word = remove_special_characters(word)
                    if len(clean_word) > 0:
                        word_index = dictonary.index(clean_word)
                        sequence.append(word_index)
                except ValueError:
                    sequence.append(unk_word_index) #Vector for unkown words
        sequences.append(sequence)
    return sequences

def run(raw_data):
    try:
        print("Received input: ", raw_data)
        
        maxSeqLength = 125
        
        print('Processing input...')
        input_data_raw = np.array(json.loads(raw_data))
        input_data_indices = convert_to_indices(input_data_raw, dictonary, contractions)
        input_data_padded = pad_sequences(input_data_indices, maxlen=maxSeqLength, padding='pre', truncating='post')
        # convert the data type to float
        input_data = np.reshape(input_data_padded.astype(np.float32), (1,maxSeqLength))
        print('Done processing input.')
        
        # Run an ONNX session to classify the input.
        result = onnx_session.run(None, {onnx_session.get_inputs()[0].name: input_data})[0].argmax(axis=1).item()
        # return just the classification index (0 or 1)
        return result
    except Exception as e:
        print(e)
        error = str(e)
        return error

## Package Model and deploy to AKS

Your scoring service can have dependencies install by using a Conda environment file. Items listed in this file will be conda or pip installed within the Docker container that is created and thus be available to your scoring web service logic.

The recommended deployment pattern is to create a deployment configuration object with the `deploy_configuration` method and then use it with the deploy method of the [Model](https://docs.microsoft.com/python/api/azureml-core/azureml.core.model.model?view=azure-ml-py) class as performed below. In this case, we use the `AksWebservice`'s `deploy_configuration` and specify the CPU cores and memory size.

The attached AKS compute name assigned to the cluster in this example is `inference-aks`. Replace with the assigned name you provided when attaching the AKS cluster within Azure Machine Learning. 

You will see output similar to the following when your web service is ready: `Succeeded - AKS service creation operation finished, operation "Succeeded"`

Run the following cell. This may take between 5-10 minutes to complete.

from azureml.core.webservice import AksWebservice, Webservice
from azureml.core.model import Model
from azureml.core.compute import AksCompute
from azureml.core.conda_dependencies import CondaDependencies
from azureml.core import Environment
from azureml.core.model import InferenceConfig

conda_packages = ['numpy==1.18.5', 'xlrd==1.2.0', 'pandas==1.0.4', 'scikit-learn==0.23.1']
pip_packages = ['azureml-defaults', 'azureml-sdk', 'tensorflow==2.2.0', 'onnxruntime==1.4.0']

environment = Environment('my-environment')
environment.python.conda_dependencies = CondaDependencies.create(conda_packages=conda_packages, 
                                                                 pip_packages=pip_packages)

execution_script = 'scoring_service.py'
service_name = "claimclassservice"

inference_config = InferenceConfig(entry_script=execution_script, environment=environment)

aks_target = AksCompute(ws,"inference-aks") #Replace with the assigned name you provided when attaching the AKS cluster within Azure Machine Learning

deployment_config = AksWebservice.deploy_configuration(cpu_cores = 1, memory_gb = 1)
service = Model.deploy(ws, service_name,[registered_model] , inference_config, deployment_config, aks_target)
service.wait_for_deployment(show_output = True)
print(service.state)
print(service.get_logs())

## Test Deployment

**Load the deployed webservice from workspace**

from azureml.core.webservice import Webservice
service_name = "claimclassservice"
service = Webservice(ws, service_name)
service

### Make direct calls on the service object

import json

test_claims = ['I crashed my car into a pole.', 
               'The flood ruined my house.', 
               'I lost control of my car and fell in the river.']

for i in range(len(test_claims)):
    result = service.run(json.dumps([test_claims[i]]))
    print('Predicted label for test claim #{} is {}'.format(i+1, result))

### Make HTTP calls to test the deployed Web Service

import requests

headers = {'Content-Type':'application/json'}

url = service.scoring_uri

for i in range(len(test_claims)):
    response = requests.post(url, json.dumps([test_claims[i]]), headers=headers)
    print('Predicted label for test claim #{} is {}'.format(i+1, response.text))

## Capture the scoring URI

In order to call the service from a REST client, you need to acquire the scoring URI. Take a note of printed scoring URI, you will need it in the last notebook.

The default settings used in deploying this service result in a service that does not require authentication, so the scoring URI is the only value you need to call this service.

print('AKS Service: Claim Classification scoring URI is: {}'.format(url))
